+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [1.12.1]
+- improvements in sprite load of the `BackgroundImageGame`.
+- improvements in `simpleAttackRangeByAngle`.
+- rename `animationTop` to `animationUp`
+- rename `animationBottom` to `animationDown`
+- improvements in `TalkDialog`
+
 # [1.12.0]
 -  Adds SpriteAnimation extension: method `asWidget`.(You can use this to SpriteAnimation or Future<SpriteAnimation>)
 -  Adds Sprite extension: method `asWidget`.(You can use this to Sprite or Future<Sprite>)